LayoutTests/fast/forms/color/display-none-input-color-chooser-shown.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/js-test.js"></script>
<input id="colorPick" type="color" />
<label for="colorPick" id="labelPick">Pick a color</label>
<script>
test (function() {
    assert_true(window.eventSender !== null);
}, "window.eventSender is required for the test to run");
test (function() {
    var input = document.createElement('input');
    input.type = 'color';
    input.value = '#000000';
    document.body.appendChild(input);

    var colorPicker = document.getElementById("colorPick");
    colorPicker.style.display = "none";
    var labelPick = document.getElementById("labelPick");
    var x = labelPick.offsetLeft + labelPick.offsetWidth/2;
    var y = labelPick.offsetTop + labelPick.offsetHeight/2;
    eventSender.mouseMoveTo(x, y);
    eventSender.mouseDown();
    eventSender.mouseUp();
    assert_true(internals.selectColorInColorChooser(input, '#ff0000'));
}, "Tests click on label for color picker with display none should show chooser");
</script>
